+++ packages/net-kit/CHANGELOG.md
@@ -1,3 +1,20 @@
+# 3.1.0
+- added `VoidModel` class for void responses
+- added `uploadMultipartData` method to upload files
+- internal refactoring
+
+## 3.0.9-dev
+
+- internal refactoring
+
+## 3.0.8-dev
+
+- added `VoidModel` class
+
+## 3.0.2-dev
+
+- added `uploadMultipartData` method
+
 # 3.0.1
 
 - updated README.md